♻️ Refactor: Quartz 트리거 설정 누락 추가
