+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/collector/reader/ArticleReader.java
@@ -8,7 +8,7 @@ import org.springframework.batch.item.ItemReader;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.collector.config.rss.RssSource;
+import com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.collector.config.batch.RssSource;
 
 import lombok.extern.slf4j.Slf4j;
