+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/collector/reader/RssSourcePartitioner.java
@@ -9,7 +9,7 @@ import org.springframework.batch.core.partition.support.Partitioner;
 import org.springframework.batch.item.ExecutionContext;
 import org.springframework.stereotype.Component;
 
-import com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.collector.config.rss.RssSource;
+import com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.collector.config.batch.RssSource;
 
 import lombok.extern.slf4j.Slf4j;
